+++ CHANGELOG.md
@@ -55,8 +55,8 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 ### Versions
 
-- Electron 19.0.0 (was 18.0.1)
-- Chromium 102.0.5005.61 (was 100.0.4896.60)
+- Electron 19.0.3 (was 18.0.1)
+- Chromium 102.0.5005.63 (was 100.0.4896.60)
 
 ## [7.2.0](https://github.com/Jelmerro/Vieb/compare/7.1.2...7.2.0) - 2022-03-31